issue_#3/DB設計
