+++ app/Models/Company.php
@@ -1,28 +1,46 @@
 <?php
 
-namespace App\Models;
+namespace App;
 
 use Illuminate\Database\Eloquent\Model;
 
 class Company extends Model
 {
-
     protected $table = 'companies';
-
+     /**
+     * The attributes that are mass assignable.
+     *
+     * @var array
+     */
     protected $fillable = [
-        'id',
-        'file_id',
-        'email',
-        'password',
+        'mission',
+        'desired_person',
         'name',
-        'description',
+        'desc',
+        'address',
+        'member_count',
+        'established_at',
+        'founder',
+        'hp_url',
+        'fb_url',
+        'tw_url',
+        'pickup_flag',
+        
     ];
 
+    public function review()
+    {
+        return $this->hasMany(\App\Models\Review::class, 'company_id', 'id');
+    }
 
-  // Relations
-  public function file()
-  {
-      return $this->belongsTo(\App\Models\File::class, 'file_id', 'id');
-  }
+    public function like()
+    {
+        return $this->hasMany(\App\Models\Like::class, 'company_id', 'id');
+    }
+
+    public function companyIndustry()
+    {
+        return $this->hasMany(\App\Models\CompanyIndustry::class, 'company_id', 'id');
+    }
 
 }
