--- app/Models/File.php
@@ -1,24 +1,0 @@
-<?php
-
-namespace App\Models;
-
-use Illuminate\Database\Eloquent\Model;
-
-class File extends Model
-{
-
-    protected $table = 'file';
-
-    protected $fillable = [
-        'url',
-        'title',
-        'entity_type',
-        's3_key',
-        's3_bucket',
-        's3_region',
-    ];
-
-
-    // Relations
-
-}
