--- app/Models/Problemset.php
@@ -1,20 +1,0 @@
-<?php
-
-namespace App\Models;
-
-use Illuminate\Database\Eloquent\Model;
-
-class Problemset extends Model
-{
-
-    protected $table = 'problemsets';
-
-    protected $fillable = [
-        'id',
-        'body',
-    ];
-
-
-  // Relations
-
-}
